Apply Qodo /improve pass 8: 64-bit widening in cap + period arithmetic
- Prevent overflow in period calculation (imp 2): FindMeasureStep computes
  periodCycles in a uint64_t intermediate so (clkFreq + freq - 1) cannot wrap.
  Real values (~100 MHz clk + <=100 kHz freq) never overflow uint32_t; this is
  defensive against a corrupted clkFreq/freq.
- Harden transport cap arithmetic (imp 1): Streaming_TransportMaxFreq widens the
  A/(B+totalChannels) divide to 64-bit so a corrupted channel count can't wrap the
  denominator. Kept the pb-init and JSON-derate comments; omitted Qodo's
  denom>0 guard since denom is always >= 2 in that branch (dead code otherwise).

Item 8 (BoardRunTimeConfig_Get NULL check) remains declined — documented CLAUDE.md
false positive, rationale on PR.

Entire-Checkpoint: 0f5b5f9b2051

+++ firmware/src/services/SCPI/SCPIInterface.c
@@ -2283,7 +2283,9 @@ static bool FindMeasureStep(StreamingRuntimeConfig* cfg, uint32_t clkFreq,
                             uint32_t wRingCap, uint32_t freq, uint32_t obsMs,
                             uint32_t* outKBps, bool* outStartFailed) {
     if (freq == 0u) { *outStartFailed = true; *outKBps = 0; return false; }  // guard div-by-zero (Qodo)
-    uint32_t periodCycles = (clkFreq + freq - 1) / freq;
+    // 64-bit intermediate so clkFreq + freq - 1 can't wrap uint32_t (Qodo pass-8
+    // hardening; the real values — ~100 MHz clk + <=100 kHz freq — never overflow).
+    uint32_t periodCycles = (uint32_t)(((uint64_t)clkFreq + freq - 1u) / freq);
     if (periodCycles < 2) periodCycles = 2;
 
     Streaming_ClearStats();
